debug

+++ src/update_v2D_half2.cl
@@ -88,7 +88,7 @@
 #endif
 
 
-extern "C" __global__ void update_v2(int a, int b ){
+extern "C" __global__ void update_v2(float a, int b ){
     
     float output;
 //    output.x = a.x+b.x;
